contrib/rtl_433_mqtt/requirements.txt
click
paho-mqtt